test: improve TS types, replace .jsx snapshot file with .tsx
