+++ components/back-to-top/PropsType.ts
@@ -1,5 +1,5 @@
-import { CSSProperties } from 'react';
-import { ContainerType } from '../utils/dom';
+import type { CSSProperties } from 'react';
+import type { ContainerType } from '../utils/dom';
 
 export default interface PropsType {
   style?: CSSProperties;
